requirements.txt
ctransformers
sentence-transformers
langchain
pinecone-client
flask
ipywidgets
langchain-pinecone
python-dotenv
-e . #to install as a local package and setup.py